fix: code review fixes for story 3-1-template-and-tool-catalog-services
- Remove redundant repository.get() call in delete() for both services
- Extract shared _AgentCatalogProtocol and _AgentRepoProtocol to services/_protocols.py
- Extract MockAgentCatalog test infrastructure to tests/helpers.py
- Add entry.id vs id consistency check in update() for both services
- Add tests for update id mismatch (2 new tests, 265 total passing)

Relates to #15
